combine project types into one file
