+++ src/renderer/context/PersistedDirectoriesContext/PersistedDirectoriesContext.ts
@@ -1,6 +1,6 @@
 import { createContext } from "react";
 import { PersistedDirectoryEntry } from "../../services/database/PersistedDirectoryEntry";
-import { Project } from "../../services/project/Project";
+import { Project } from "../../services/project/types";
 
 interface PersistedDirectoriesContextProps {
   checkWorkingDirectoryPermission: () => Promise<void>;
