--- src/renderer/services/project/Project.ts
@@ -1,9 +1,0 @@
-import { FileInfoId, IsoDateTimeString } from "../Flavors";
-
-export interface Project {
-  name: string;
-  directoryName: string;
-  projectFrameRate: number;
-  lastSaved: IsoDateTimeString;
-  fileInfoId: FileInfoId;
-}
