--- src/renderer/services/project/Track.ts
@@ -1,9 +1,0 @@
-import { FileInfoType } from "../fileManager/FileInfo";
-import { TrackId } from "../Flavors";
-import { TrackItem } from "./TrackItem";
-
-export interface Track {
-  id: TrackId;
-  fileType: FileInfoType;
-  trackItems: TrackItem[];
-}
