--- src/renderer/services/project/TrackItem.ts
@@ -1,10 +1,0 @@
-import { FileInfoId, FrameCount, TimelineIndex, TrackGroupId, TrackItemId } from "../Flavors";
-
-export interface TrackItem {
-  id: TrackItemId;
-  length: FrameCount;
-  fileName: string;
-  fileNumber: TimelineIndex;
-  trackGroupId: TrackGroupId;
-  fileInfoId: FileInfoId;
-}
