+++ src/renderer/services/utils.ts
@@ -17,3 +17,4 @@ export const DEFAULT_PROJECT_DIRECTORY_NAME = `Untitled-Movie.${PROJECT_DIRECTOR
 export const DEFAULT_PROJECT_FRAME_RATE = 15;
 export const DEFAULT_ONION_SKIN_OPACITY = 0.5;
 export const DEFAULT_ONION_SKIN_FRAMES_VISIBLE = 1;
+export const CURRENT_PROJECT_INFO_FILE_SCHEMA_VERSION = 1;
